+++ Common/src/main/java/com/tristankechlo/crop_marker/mixin/MultiVariantMixin.java
@@ -1,6 +1,5 @@
 package com.tristankechlo.crop_marker.mixin;
 
-import com.tristankechlo.crop_marker.types.MarkerPosition;
 import com.tristankechlo.crop_marker.util.ResourceLocationHelper;
 import net.minecraft.client.renderer.block.model.MultiVariant;
 import net.minecraft.client.renderer.block.model.Variant;
@@ -25,14 +24,15 @@ public abstract class MultiVariantMixin {
 
     @Inject(method = "bake", at = @At("HEAD"))
     private void FullGrownCropMarker$bake(ModelBaker $$0, Function<Material, TextureAtlasSprite> $$1, ModelState $$2, ResourceLocation modelId, CallbackInfoReturnable<@Nullable BakedModel> cir) {
-        MarkerPosition marker = MarkerPosition.fromId(modelId);
-        if (marker != MarkerPosition.NONE) {
+        boolean shouldHaveMarker = ResourceLocationHelper.FullGrownCropMarker$shouldHaveMarker(modelId);
+        if (shouldHaveMarker) {
             this.getVariants().forEach(variant -> {
                 //reapply the marker to all the variants of this model
                 //this is needed because the models are loaded as a WeightedUnbakedModel from the blockstates json
                 //the WeightedUnbakedModel then loads the variant models from the actual models json
                 ResourceLocationHelper location = (ResourceLocationHelper) variant.getModelLocation();
-                location.FullGrownCropMarker$setMarkerPosition(((ResourceLocationHelper) modelId).FullGrownCropMarker$getMarkerPosition());
+                boolean hasMarker = location.FullGrownCropMarker$shouldHaveMarker();
+                location.FullGrownCropMarker$setShouldHaveMarker(hasMarker);
             });
         }
     }