+++ Common/src/main/java/com/tristankechlo/crop_marker/mixin/ResourceLocationMixin.java
@@ -1,6 +1,5 @@
 package com.tristankechlo.crop_marker.mixin;
 
-import com.tristankechlo.crop_marker.types.MarkerPosition;
 import com.tristankechlo.crop_marker.util.ResourceLocationHelper;
 import net.minecraft.resources.ResourceLocation;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,16 +7,16 @@ import org.spongepowered.asm.mixin.Mixin;
 @Mixin(ResourceLocation.class)
 public abstract class ResourceLocationMixin implements ResourceLocationHelper {
 
-    protected MarkerPosition FullGrownCropMarker$markerPosition = MarkerPosition.NONE;
+    protected boolean FullGrownCropMarker$shouldHaveMarker = false;
 
     @Override
-    public MarkerPosition FullGrownCropMarker$getMarkerPosition() {
-        return FullGrownCropMarker$markerPosition;
+    public void FullGrownCropMarker$setShouldHaveMarker(boolean hasMarker) {
+        this.FullGrownCropMarker$shouldHaveMarker = hasMarker;
     }
 
     @Override
-    public void FullGrownCropMarker$setMarkerPosition(MarkerPosition position) {
-        FullGrownCropMarker$markerPosition = position;
+    public boolean FullGrownCropMarker$shouldHaveMarker() {
+        return this.FullGrownCropMarker$shouldHaveMarker;
     }
 
 }