+++ Changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### Version 1.19.3 - 1.1.0
+
+- add support for `minecraft:sweet_berry_bush`
+- rename config option to disable the marker
+    - previous config option `position` is no longer supported
+    - new config option `hasMarker` is now used
+
 ### Version 1.19.3 - 1.0.0
 
 - initial release